Update Heartbeat to typescript
